Extract code related to the collection of real errors from real.py to collect_real_errors.py, and make it executable
