+++ python/git_helper.py
@@ -1,10 +1,12 @@
 import git
 from git import Repo
+from git import InvalidGitRepositoryError
 import os
 import configparser
 
+dir_path = os.path.dirname(os.path.realpath(__file__))
 config = configparser.ConfigParser()
-config.read('./config.ini')
+config.read(os.path.join(dir_path, "config.ini"))
 
 __git_repo_dir = config['DEFAULT']['git_repo_dir']
 
@@ -14,7 +16,7 @@ def clone_repo(user, repo_name):
     Clone the repo into the repo_dir location
     """
     dir = get_repo_dir(user, repo_name)
-    print(f'Clonning {user}/{repo_name}')
+    print(f'Cloning {user}/{repo_name}')
     return Repo.clone_from(f'git@github.com:{user}/{repo_name}.git', dir)
 
 
@@ -31,7 +33,11 @@ def open_repo(user, repo_name):
     """
     dir = get_repo_dir(user, repo_name)
     if os.path.exists(dir):
-        return Repo(dir)
+        try:        
+            return Repo(dir)
+        except InvalidGitRepositoryError:
+            print("Repo %s not found." % dir)
+            return None
     else:
         return clone_repo(user, repo_name)
